Anirud/re update license headers(#51)
* update lc header python file
- to use pathlib instead of os for compatability
simalr to https://github.com/tenstorrent/tt-inference-server/pull/8/commits/00b5764adca4f36e38e7c01a2abb87b8ec1015a2
- also work on backend and models folder

* add any other missing lc headers
