+++ models/dummy_echo_model/src/dummy_echo_backend.py
@@ -1,3 +1,7 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
 import time
 import traceback
 from multiprocessing import Queue
