+++ models/tt-metal-falcon-7b/src/_mock_inference_api_server.py
@@ -1,3 +1,7 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
 import os
 from time import sleep
 from unittest.mock import Mock, patch
