+++ models/tt-metal-falcon-7b/src/inference_config.py
@@ -1,3 +1,7 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
 import os
 from collections import namedtuple
 from pprint import pprint
