+++ models/tt-metal-falcon-7b/src/inference_logger.py
@@ -1,3 +1,7 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
 import logging
 import os
 from datetime import datetime
